pyproject.toml

[project]
name = "pyarpg"
version = "0.1.0"
description = "An ARPG made with pygame."
requires-python = "==3.13.*"
dependencies = [
  "pygame-ce>=2.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pyarpg"]

[dependency-groups]
dev = [
  "ruff",
]

[tool.ruff]
line-length = 88
target-version = "py313"